Prevents vim swap files to be found as rules
Exception testing searched for all files inside the rules directory and
tried to come up with a class name to it. VIM swap files always start
with a dot (since they are hidden files) and fail to be a valid class
name, provoking an error while testing.

Redesign the tests to use that "rule list" as a data provider instead
of being an algorithmn inside a test. Getting rid of the iteration inside
the test we can test all exception properties more easily while being
able to fail just one specific test (for an especific exception rule)
instead of preventing any more tests on exceptions from being executed.

While the test has fewer lines now, coverage and asserts are the same as
before.